==> sick leaves/01 modelos caida lme sti.Rmd <==
---
title: "Modelos de serie de tiempo interrumpida para caída en lme oncológicas"
output: 
---

```{r RMD SETTINGS, message=FALSE, warning=FALSE, include=FALSE}
rm(list = ls()); gc() 
knitr::opts_knit$set(root.dir = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme")
knitr::opts_chunk$set(include = FALSE,
                      echo = FALSE,
                      message = FALSE,
                      warning = FALSE)
```

```{r LIBRERIAS, message=FALSE, warning=FALSE, include=FALSE}
library(data.table)
library(tidyverse)
library(MASS)
library(sjPlot)
```

# 1. Datos

```{r LOAD AND PREPARE SAVED DATA GROUPED BY WEEK}
load("output/DB/licencias_cancer_2018.RData")

lme_modelos_agregados <- licencias_cancer_2018 %>% 
  # FILTRO
  filter(prevision %in% c("FONASA", "ISAPRE"), !is.na(grupo_edad_2), 
         !is.na(sexo_trabajador))  %>% 
  group_by(date_wk, cie10, prevision, sexo_trabajador, grupo_edad_2) %>% 
  summarise(n_lme=sum(n_lme))

lme.2018.dt <- as.data.table(lme_modelos_agregados)

lme_modelos_agregados <- lme.2018.dt[
  , time := as.numeric(factor(date_wk, labels = 0:156)) - 1  ][
    , time_after := ifelse(time < 115, 0, (time - 115 ))  ]

rm(lme.2018.dt)

lme_modelos_agregados <- lme_modelos_agregados  %>% 
  # VARIABLES NUEVAS Y MODIFICADAS
  mutate(pandemia = ifelse(date_wk < as.Date("2020-03-15"), 0, 1),
         grupo_edad_2 = relevel(grupo_edad_2, "[40,50)"),
         sexo_trabajador = as.factor(sexo_trabajador),
         sexo_trabajador = relevel(sexo_trabajador, "M"),
         year=lubridate::year(date_wk),
         week=lubridate::week(date_wk),
         prevision = as.factor(prevision),
         week = as.factor(week)) %>% 
  rename(grupo_edad = grupo_edad_2)

lme_modelos_agregados <- droplevels(lme_modelos_agregados)

sum(lme_modelos_agregados$n_lme)

cancer_sexo <- cbind("C50","C51","C52","C53","C54","C55","C56","C57","C58", 
                     "C60","C61","C62","C63")  
```


```{r MODELO AGREGADO ADITIVO Y POR SUBSETS}

dat.ag <- lme_modelos_agregados %>% 
  rename(edad_cat = grupo_edad,
         sexo = sexo_trabajador, 
         count = n_lme)
# Agregado y por subsets

sets <- list(
  dat.ag,
  dat.ag[!(cie10 %in% cancer_sexo)][sexo == "F"],
  dat.ag[!(cie10 %in% cancer_sexo)][sexo == "M"],
  dat.ag[prevision == "ISAPRE"],
  dat.ag[prevision == "FONASA"],
#  dat.ag[edad_cat == "[0,10)"],
#  dat.ag[edad_cat == "[10,20)"],
  dat.ag[edad_cat == "[20,30)"],
  dat.ag[edad_cat == "[30,40)"],
  dat.ag[edad_cat == "[40,50)"],
  dat.ag[edad_cat == "[50,60)"],
  dat.ag[edad_cat == "[60,70)"],
  dat.ag[edad_cat == "[70,80)"])

formulas <- list(
  as.formula(count ~ sexo + edad_cat + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ edad_cat + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ edad_cat + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + edad_cat + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + edad_cat + year + week + pandemia + time + pandemia:time_after),
#  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
# as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after))

models <- list()

for (i in 1:11){
  
  sets[[i]] <- sets[[i]] %>% group_by(date_wk, prevision, sexo, edad_cat, year, 
                                      week, time, pandemia, time_after ) %>% 
  summarise(count=sum(count))
  
  models[[i]] <- glm.nb(formulas[[i]], data = sets[[i]])
  
}

save(models, file = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/modelos/nb sti caida lme/20210221 sti caida lme agregados.RData" )

#load(file = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/modelos/nb sti caida lme/20210221 sti caida lme agregados.RData" )
```

## 2.1 Tablas modelos

```{r}

tab.models <- data.table()

for (i in 1:11){

  summ <- summary(models[[i]])
  mod.res <- as.data.frame(summ$coefficients)
  mod.res$Model <- i
  mod.res$Vars <- rownames(summ$coefficients)

  tab.models <- rbind(tab.models, mod.res)
  
}

rm(summ, mod.res, i)

confint.models <- data.table()

for (i in 1:11){
  
  conf <- confint(models[[i]])
  conf.res <- as.data.frame(conf)
  conf.res$Model <- i
  conf.res$Vars <- rownames(conf)

  confint.models <- rbind(confint.models, conf.res)
  
}

rm(conf, conf.res, i)

tab.models <- cbind(tab.models, confint.models[,c(1:2)])

invlink <- models[[1]]$family$linkinv

. <- tab.models[
  , IRR := invlink(Estimate) ][
  , IC_lower := invlink(`2.5 %`) ][
  , IC_upper := invlink(`97.5 %`) ]

tab.models <- .

save(tab.models, file = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/tab/tab nb sti/20210223 coef sti datos agregados.RData")

#load(file = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/tab/tab nb sti/20210223 coef sti datos agregados.RData")
```


## 2.2 Gráfico de coeficientes de modelos

```{r fig.height=10, fig.width=6}
tab.models %>% 
  ggplot(aes(x = IRR, y = Vars)) +
  geom_point()
```

```{r}
g.coeff <- tab.models %>% 
  filter(Vars %in% c("(Intercept)", "edad_cat[0,10)", "edad_cat[10,20)", "edad_cat[20,30)", 
                       "edad_cat[30,40)", "edad_cat[50,60)", "edad_cat[60,70)", "edad_cat[70,80)", 
                       "pandemia", "pandemia:time_after", "previsionISAPRE", "sexoF", "time")) %>% 
 mutate(Mod.int = as.factor(Model),
         Mod.int = factor(Mod.int, labels = c("Additive", rep("Sex", 2), rep("Insurance", 2), rep("Age", 6))),
         Model = as.factor(Model),
         Model = factor(Model, labels = c("Additive", "Female", "Male", "Private", "Public", 
                                          "[20,30)", "[30,40)", "[40,50)", "[50,60)", "[60,70)", "[70,80)"))) %>% 
  
  ggplot(aes(x = IRR, y = Vars)) +
  geom_point(aes(color = Model)) +
  geom_errorbarh(aes(xmin = IC_lower, xmax = IC_upper, color = Model), alpha = .25) +
  geom_vline(xintercept = 1, alpha = .3, linetype = 2) +
  facet_grid(. ~ Mod.int, scales = "free_x")

g.coeff

ggsave(g.coeff, 
       filename = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/graf/2020/coef modelos.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

```

# 3. Predicciones

```{r PREDICTED DB}
dat.plots.list <- list()

for (i in 1:11) {
  
  dat.cf <- sets[[i]] 
  dat.cf$pandemia <- 0

  dat.plots.temp <- sets[[i]]

  ilink <- family(models[[i]])$linkinv
  
  pred <- predict(models[[i]], se.fit = T)
  yhat <- pred$fit
  se <- pred$se.fit
  
  dat.plots.temp$yhat <- ilink(yhat)
  dat.plots.temp$ub <- ilink(yhat + (1.96*se))
  dat.plots.temp$lb <- ilink(yhat - (1.96*se))
  
  
  pred.cf <- predict(models[[i]], newdata = dat.cf, se.fit = T)
  cf <- pred.cf$fit
  cf.se <- pred.cf$se.fit
  
  dat.plots.temp$cf <- ilink(cf)
  dat.plots.temp$cf_ub <- ilink(cf + (1.96*cf.se))
  dat.plots.temp$cf_lb <- ilink(cf - (1.96*cf.se))
  
  dat.plots.list[[i]] <- dat.plots.temp
  
  rm(dat.cf, dat.plots.temp, ilink, pred, yhat, se, pred.cf, cf, cf.se)
}

```

```{r RECODE AND BIND PREDICTED DB}
mod <- c("Additive", 
        rep("Sex", 2), 
         rep("Insurance", 2), 
         rep("Age", 6))

for (i in 1:11) {

#  levels(dat.plots.list[[i]]$grp.cancer) <- c("General Oncological Services", "Colorectal Cancer", "Cervical Cancer",
#                                  "Gastric Cancer", "Breast Cancer", "Prostate Cancer")

#  levels(dat.plots.list[[i]]$grupo.trz) <- c("Outpatient Visit", "Antigens\n(Ca 125, Ca 15-3, Ca 19-9, Cea)", 
#                                   "Biopsies",   "Neoplastic Cells" , "PET-CT", 
#                                 "FOBT", "Sigmoidoscopy and Colonoscopy", 
#                                 "Cervical Biopsy", "Colposcopy", "Smear (PAP) test", 
#                                "Upper GI endoscopy ", "Urease test", 
#                                 "Breast US", "Mammography", "Breast MRI",
#                                 "PSA", "Prostate US")

  dat.plots.list[[i]]$mod.int <- mod[i]
  
}

dat.plots <- data.table()

for (i in 1:11){
  
  dat.plots <- rbind(dat.plots, dat.plots.list[[i]])
  
}

rm(dat.plots.list)

save(dat.plots, file = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/DB/20210223 DB predichos nb sti.RData")
#load("/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/DB/20210223 DB predichos nb sti.RData")
```

# 4. Plots

## 4.1 Serie completa

```{r}
g1.a <- dat.plots %>% 
  filter(mod.int == "Additive") %>% 
  group_by(date_wk) %>% summarise(count = sum(count),
                                  yhat = sum(yhat),
                                  cf = sum(cf),
                                  ub = sum(ub),
                                  lb = sum(lb),
                                  cf_ub = sum(cf_ub),
                                  cf_lb = sum(cf_lb)) %>% 
  ggplot(aes(x = date_wk, y = count)) +
  geom_point(alpha = .3) +
  geom_line(aes(x = date_wk, y = yhat)) +
  geom_ribbon(aes(ymin = lb, ymax = ub), alpha = .3) +
  geom_line(aes(x = date_wk, y = cf), linetype = 2, size = .3) +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub), alpha = .15) +
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Date", 
       title = "Additive model") +
  geom_vline(xintercept = as.Date("2020-03-15"), linetype = "dotted", color = "red") +
  theme(legend.position = "bottom",
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 9),
        plot.subtitle = element_text(size = 14),
        plot.title = element_text(size = 12),
        strip.text = element_text(size = 14))
  
g1.a

ggsave(g1.a, 
       filename = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/graf/2020/serie completa adit.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

```

```{r}
g1.b <- dat.plots %>% 
  filter(mod.int == "Sex") %>%
  mutate(sexo = fct_relevel(sexo, sort)) %>% 
  group_by(date_wk, sexo) %>% summarise(count = sum(count),
                                  yhat = sum(yhat),
                                  cf = sum(cf),
                                  ub = sum(ub),
                                  lb = sum(lb),
                                  cf_ub = sum(cf_ub),
                                  cf_lb = sum(cf_lb)) %>% 
  ggplot(aes(x = date_wk, y = count)) +
  geom_point(alpha = .3, aes(color = sexo)) +
  geom_line(aes(x = date_wk, y = yhat, color = sexo)) +
  geom_ribbon(aes(ymin = lb, ymax = ub, fill = sexo), alpha = .3) +
  geom_line(aes(x = date_wk, y = cf, color = sexo), linetype = 2, size = .3) +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, fill = sexo), alpha = .15) +
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Date", fill = "Sexo", color = "Sexo",
       title = "by sex") +
  geom_vline(xintercept = as.Date("2020-03-15"), linetype = "dotted", color = "red") +
  theme(legend.position = "bottom",
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 9),
        plot.subtitle = element_text(size = 14),
        plot.title = element_text(size = 12),
        strip.text = element_text(size = 14))
  
g1.b

ggsave(g1.b, 
       filename = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/graf/2020/serie completa sexo.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)
```

```{r}
g1.c <- dat.plots %>% 
  filter(mod.int == "Insurance") %>% 
  group_by(date_wk, prevision) %>% summarise(count = sum(count),
                                  yhat = sum(yhat),
                                  cf = sum(cf),
                                  ub = sum(ub),
                                  lb = sum(lb),
                                  cf_ub = sum(cf_ub),
                                  cf_lb = sum(cf_lb)) %>% 
  ggplot(aes(x = date_wk, y = count)) +
  geom_point(aes(color = prevision), alpha = .3) +
  geom_line(aes(x = date_wk, y = yhat, color = prevision)) +
  geom_ribbon(aes(ymin = lb, ymax = ub, fill = prevision), alpha = .3) +
  geom_line(aes(x = date_wk, y = cf, color = prevision), linetype = 2, size = .3) +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, fill = prevision), alpha = .15) +
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Date", fill = "Previsión", color = "Previsión",
       title = "by insurance") +
  geom_vline(xintercept = as.Date("2020-03-15"), linetype = "dotted", color = "red") +
  theme(legend.position = "bottom",
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 9),
        plot.subtitle = element_text(size = 14),
        plot.title = element_text(size = 12),
        strip.text = element_text(size = 14))
  
g1.c

ggsave(g1.c, 
       filename = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/graf/2020/serie completa prevision.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)
```

```{r}
g1.d <- dat.plots %>% 
  filter(mod.int == "Age") %>% 
  mutate(edad_cat = fct_relevel(edad_cat, sort) ) %>% 
  group_by(date_wk, edad_cat) %>% summarise(count = sum(count),
                                  yhat = sum(yhat),
                                  cf = sum(cf),
                                  ub = sum(ub),
                                  lb = sum(lb),
                                  cf_ub = sum(cf_ub),
                                  cf_lb = sum(cf_lb)) %>% 
  ggplot(aes(x = date_wk, y = count)) +
  geom_point(aes(color = edad_cat), alpha = .15) +
  geom_line(aes(x = date_wk, y = yhat, color = edad_cat)) +
  geom_ribbon(aes(ymin = lb, ymax = ub, fill = edad_cat), alpha = .3) +
  geom_line(aes(x = date_wk, y = cf, color = edad_cat), linetype = 2, size = .3) +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, fill = edad_cat), alpha = .075) +
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Date", fill = "Edad", color = "Edad",
       title = "by age") +
  geom_vline(xintercept = as.Date("2020-03-15"), linetype = "dotted", color = "red") +
  theme(legend.position = "bottom",
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 9),
        plot.subtitle = element_text(size = 14),
        plot.title = element_text(size = 12),
        strip.text = element_text(size = 14))
  
g1.d

ggsave(g1.d, 
       filename = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/graf/2020/serie completa edad.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)
```

## 4.2 Exposición

```{r}
g2.a <- dat.plots %>% 
  mutate(week = as.numeric(week)) %>% 
  filter(mod.int == "Additive") %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019"),
         cf = ifelse(exposicion == "2020" & pandemia != 0, cf, NA),
         cf_ub = ifelse(exposicion == "2020" & pandemia != 0, cf_ub, NA),
         cf_lb = ifelse(exposicion == "2020" & pandemia != 0, cf_lb, NA) ) %>% 
  group_by(week, year, exposicion) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion)%>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = sum(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb) ) %>% 
  
  ggplot(aes(x = week, y = count, fill = exposicion))+
  
  geom_point(alpha = .3, aes(color = exposicion), size = 1.5) +
  geom_line(aes(x = week, y = yhat, color = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +

  geom_line(aes(x = week, y = cf, color = exposicion, linetype = "Predicted counterfactual"), 
            size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       title = "Sick leave due to Cancer") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 10),
        plot.title = element_text(size = 12))

g2.a

ggsave(g2.a, 
       filename = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/graf/2020/2020 vs anterior aditivo.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

```

```{r}
g2.b <- dat.plots %>% 
  mutate(week = as.numeric(week)) %>% 
  filter(mod.int == "Sex") %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019"),
         cf = ifelse(exposicion == "2020" & pandemia != 0, cf, NA),
         cf_ub = ifelse(exposicion == "2020" & pandemia != 0, cf_ub, NA),
         cf_lb = ifelse(exposicion == "2020" & pandemia != 0, cf_lb, NA)) %>% 
  group_by(week, year, exposicion, sexo) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion, sexo)%>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = sum(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb) ) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  
  geom_point(alpha = .3, aes(color = exposicion), size = 1.5) +
  geom_line(aes(x = week, y = yhat, color = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +

  geom_line(aes(x = week, y = cf, color = exposicion, linetype = "Predicted counterfactual"), size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  
  facet_grid(. ~ sexo) + 
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       title = "Sick leaves due to Cancer",
       subtitle = "by sex") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 10),
        plot.title = element_text(size = 12))

g2.b

ggsave(g2.b, 
       filename = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/graf/2020/2020 vs anterior sexo.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

```

```{r}
g2.c <- dat.plots %>% 
  mutate(week = as.numeric(week)) %>% 
  filter(mod.int == "Insurance") %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019"),
         cf = ifelse(exposicion == "2020" & pandemia != 0, cf, NA),
         cf_ub = ifelse(exposicion == "2020" & pandemia != 0, cf_ub, NA),
         cf_lb = ifelse(exposicion == "2020" & pandemia != 0, cf_lb, NA)) %>% 
  group_by(week, year, exposicion, prevision) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion, prevision)%>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = sum(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb) ) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  
  geom_point(alpha = .3, aes(color = exposicion), size = 1.5) +
  geom_line(aes(x = week, y = yhat, color = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +

  geom_line(aes(x = week, y = cf, color = exposicion, linetype = "Predicted counterfactual"), size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  
  facet_grid(. ~ prevision) + 
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       title = "Sick leaves due to Cancer",
       subtitle = "by insurance") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 10),
        plot.title = element_text(size = 12))

g2.c

ggsave(g2.c, 
       filename = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/graf/2020/2020 vs anterior prevision.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)
```

```{r}
g2.d <- dat.plots %>% 
  mutate(week = as.numeric(week)) %>% 
  filter(mod.int == "Age") %>% 
  mutate(edad_cat = fct_relevel(edad_cat, sort),
         exposicion = ifelse(year == 2020, "2020", "2018-2019"),
         cf = ifelse(exposicion == "2020" & pandemia != 0, cf, NA),
         cf_ub = ifelse(exposicion == "2020" & pandemia != 0, cf_ub, NA),
         cf_lb = ifelse(exposicion == "2020" & pandemia != 0, cf_lb, NA)) %>% 
  group_by(week, year, exposicion, edad_cat) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion, edad_cat)%>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = sum(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb) ) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  
  geom_point(alpha = .3, aes(color = exposicion), size = 1.5) +
  geom_line(aes(x = week, y = yhat, color = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +

  geom_line(aes(x = week, y = cf, color = exposicion, linetype = "Predicted counterfactual"), size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  
  facet_wrap("edad_cat", nrow = 2) + 
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       title = "Sick leaves due to Cancer",
       subtitle = "by age group") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 10),
        plot.title = element_text(size = 12))

g2.d

ggsave(g2.d, 
       filename = "/Users/Fran/Dropbox (Personal)/Analisis IMED/lme/output/graf/2020/2020 vs anterior edad.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

```
